+++ x/vpool/keeper/keeper_test.go
@@ -174,6 +174,7 @@ func TestSwapQuoteForBase(t *testing.T) {
 					MaintenanceMarginRatio: sdk.MustNewDecFromStr("0.0625"),
 					MaxLeverage:            sdk.MustNewDecFromStr("15"),
 				},
+				sdk.ZeroDec(),
 			))
 			vpool, err := vpoolKeeper.GetPool(ctx, asset.Registry.Pair(denoms.BTC, denoms.NUSD))
 			require.NoError(t, err)
@@ -360,6 +361,7 @@ func TestSwapBaseForQuote(t *testing.T) {
 					MaintenanceMarginRatio: sdk.MustNewDecFromStr("0.0625"),
 					MaxLeverage:            sdk.MustNewDecFromStr("15"),
 				},
+				sdk.ZeroDec(),
 			))
 
 			vpool, err := vpoolKeeper.GetPool(ctx, asset.Registry.Pair(denoms.BTC, denoms.NUSD))
@@ -407,6 +409,7 @@ func TestGetVpools(t *testing.T) {
 			MaintenanceMarginRatio: sdk.MustNewDecFromStr("0.0625"),
 			MaxLeverage:            sdk.MustNewDecFromStr("15"),
 		},
+		sdk.ZeroDec(),
 	))
 	assert.NoError(t, vpoolKeeper.CreatePool(
 		ctx,
@@ -420,6 +423,7 @@ func TestGetVpools(t *testing.T) {
 			MaintenanceMarginRatio: sdk.MustNewDecFromStr("0.0625"),
 			MaxLeverage:            sdk.MustNewDecFromStr("15"),
 		},
+		sdk.ZeroDec(),
 	))
 
 	pools := vpoolKeeper.Pools.Iterate(ctx, collections.Range[asset.Pair]{}).Values()
@@ -437,6 +441,7 @@ func TestGetVpools(t *testing.T) {
 			MaintenanceMarginRatio: sdk.MustNewDecFromStr("0.0625"),
 			MaxLeverage:            sdk.MustNewDecFromStr("15"),
 		},
+		Bias: sdk.ZeroDec(),
 	}))
 	require.EqualValues(t, pools[1], types.NewVpool(types.ArgsNewVpool{
 		Pair:          asset.Registry.Pair(denoms.ETH, denoms.NUSD),
@@ -449,6 +454,7 @@ func TestGetVpools(t *testing.T) {
 			MaintenanceMarginRatio: sdk.MustNewDecFromStr("0.0625"),
 			MaxLeverage:            sdk.MustNewDecFromStr("15"),
 		},
+		Bias: sdk.ZeroDec(),
 	}))
 }